requirements.txt
flask>=3.1.0
flask-wtf>=1.2.0
gunicorn>=23.0.0
nh3>=0.2.0
supabase>=2.24.0
requests>=2.31.0
python-dotenv>=1.0.0
